+++ Krowi_AchievementFilter/Krowi_AchievementFilter.lua
@@ -1,10 +1,7 @@
-AF_COLOR = "|cFF1D92C2";
-AF_NAME = "Krowi's " .. AF_COLOR .. "Achievement Filter|r";
-AF_VERSION = GetBuildInfo();
-AF_BUILD = GetAddOnMetadata("Krowi_AchievementFilter", "Version");
-AF_VERSION_BUILD = AF_VERSION .. "." .. AF_BUILD;
+local addonName, addon = ...;
 
-local _, addon = ...; -- Global addon namespace
+addon.L = LibStub(addon.Libs.AceLocale):GetLocale(addonName);
+
 addon.Faction = {}; -- Global faction data
 addon.Faction.IsAlliance = UnitFactionGroup("player") == "Alliance";
 addon.Faction.IsHorde = UnitFactionGroup("player") == "Horde";
@@ -24,7 +21,7 @@ function loadHelper:OnEvent(event, arg1)
             addon.Diagnostics.Load();
             addon.Options.Load();
             addon.Icon.Load();
-
+            addon.Tutorials.Load();
         elseif arg1 == "Blizzard_AchievementUI" then -- This needs the Blizzard_AchievementUI addon available to load
             local achievementsFrame = addon.GUI.AchievementsFrame:New();
             local categoriesFrame = addon.GUI.CategoriesFrame:New(addon.Categories, achievementsFrame.Frame);