+++ Krowi_AchievementFilter/Libs/Libs.lua
@@ -1,0 +1,6 @@
+local _, addon = ...; -- Global addon namespace
+addon.Libs = {}; -- Global library names table
+local libs = addon.Libs; -- Local library names table
+
+libs.AceLocale = "AceLocale-3.0";
+libs.Tutorials = "MSA-Tutorials-1.0";